Modify Auth logic

+++ backend/api/routers/post.py
@@ -1,7 +1,7 @@
 import os
 from typing import List
 import api.schemas.post as post_schema
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from google.cloud import firestore
 from .auth import verify_token
 
@@ -11,8 +11,10 @@ db = firestore.Client(os.getenv("GOOGLE_CLOUD_PROJECT"))
 
 @router.get("/posts", response_model=List[post_schema.Post])
 async def get_root():
+    # 投稿を取得
     docs = db.collection(u"posts").order_by(
         u"timestamp", direction=firestore.Query.DESCENDING).get()
+    # 投稿データを整形
     posts = []
     for doc in docs:
         data = doc.to_dict()
@@ -20,21 +22,32 @@ async def get_root():
         posts.append(data)
     return posts
 
-@router.post("/posts", response_model=post_schema.PostCreateResponse)
-async def post_root(post_body: post_schema.PostCreate, info = Depends(verify_token)):
-    print(info)
+
+@router.post("/posts")
+async def post_root(post_body: post_schema.PostCreate, id_info = Depends(verify_token)):
+    # 投稿データ
     data = {
         u"timestamp": firestore.SERVER_TIMESTAMP,
-        u"user_name": post_body.user_name,
-        u"user_email": post_body.user_email,
-        u"user_image": post_body.user_image,
+        u"user_name": id_info["name"],
+        u"user_email": id_info["email"],
+        u"user_image": id_info["picture"],
         u"text": post_body.text
     }
+    # 投稿を作成
     db.collection(u"posts").document().set(data)
-    return post_schema.PostCreateResponse(**post_body.dict())
+    return {"message": "Post created successfully"}
 
 
 @router.delete("/posts")
-async def delete_root(post_delete_body: post_schema.PostDelete, _: bool = Depends(verify_token)):
+async def delete_root(post_delete_body: post_schema.PostDelete, id_info = Depends(verify_token)):
+    doc = db.collection(u"posts").document(post_delete_body.id).get()
+    # 存在しない投稿を削除しようとした場合
+    if not doc.exists:
+        raise HTTPException(status_code=404, detail="Post not found")
+    doc_email = doc.to_dict()["user_email"]
+    # 他人の投稿を削除しようとした場合
+    if doc_email != id_info["email"]:
+        raise HTTPException(status_code=403, detail="Forbidden")
+    # 投稿を削除
     db.collection(u"posts").document(post_delete_body.id).delete()
     return {"message": "Post deleted successfully"}
